Fix PR remarks and refactoring
