+++ hedera-mirror-test/src/test/java/com/hedera/mirror/test/e2e/acceptance/steps/ContractFeature.java
@@ -47,7 +47,6 @@ import io.cucumber.java.en.When;
 import java.math.BigInteger;
 import java.util.Comparator;
 import java.util.HexFormat;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.io.Resource;
@@ -56,7 +55,7 @@ import org.springframework.util.StringUtils;
 import org.springframework.web.client.HttpClientErrorException;
 
 @RequiredArgsConstructor
-public class ContractFeature extends AbstractFeature {
+public class ContractFeature extends BaseContractFeature {
     private static final String GET_ACCOUNT_BALANCE_SELECTOR = "6896fabf";
     private static final String GET_SENDER_SELECTOR = "5e01eb5a";
     private static final String MULTIPLY_SIMPLE_NUMBERS_SELECTOR = "8070450f";
@@ -67,7 +66,7 @@ public class ContractFeature extends AbstractFeature {
     private final AccountClient accountClient;
     private final MirrorNodeClient mirrorClient;
     private final Web3Properties web3Properties;
-    private DeployedContract deployedParentContract;
+
     private String create2ChildContractEvmAddress;
     private String create2ChildContractEntityId;
     private AccountId create2ChildContractAccountId;
@@ -305,11 +304,9 @@ public class ContractFeature extends AbstractFeature {
         executeSelfDestructTransaction();
     }
 
-    private ContractResponse verifyContractFromMirror(boolean isDeleted) {
-        var mirrorContract =
-                mirrorClient.getContractInfo(deployedParentContract.contractId().toString());
+    protected ContractResponse verifyContractFromMirror(boolean isDeleted) {
+        var mirrorContract = super.verifyContractFromMirror(isDeleted);
 
-        assertNotNull(mirrorContract);
         assertThat(mirrorContract.getAdminKey()).isNotNull();
         assertThat(mirrorContract.getAdminKey().getKey())
                 .isEqualTo(contractClient
@@ -317,93 +314,28 @@ public class ContractFeature extends AbstractFeature {
                         .getExpandedOperatorAccountId()
                         .getPublicKey()
                         .toStringRaw());
-        assertThat(mirrorContract.getAutoRenewPeriod()).isNotNull();
-        assertThat(mirrorContract.getBytecode()).isNotBlank();
-        assertThat(mirrorContract.getContractId())
-                .isEqualTo(deployedParentContract.contractId().toString());
-        assertThat(mirrorContract.getCreatedTimestamp()).isNotBlank();
-        assertThat(mirrorContract.getDeleted()).isEqualTo(isDeleted);
-        assertThat(mirrorContract.getFileId())
-                .isEqualTo(deployedParentContract.fileId().toString());
-        assertThat(mirrorContract.getMemo()).isNotBlank();
-        String address = mirrorContract.getEvmAddress();
-        assertThat(address).isNotBlank().isNotEqualTo("0x").isNotEqualTo("0x0000000000000000000000000000000000000000");
-        assertThat(mirrorContract.getTimestamp()).isNotNull();
-        assertThat(mirrorContract.getTimestamp().getFrom()).isNotNull();
-
-        if (contractClient
-                .getSdkClient()
-                .getAcceptanceTestProperties()
-                .getFeatureProperties()
-                .isSidecars()) {
-            assertThat(mirrorContract.getRuntimeBytecode()).isNotNull();
-        }
-
-        assertThat(mirrorContract.getBytecode())
-                .isEqualTo(deployedParentContract.compiledSolidityArtifact().getBytecode());
-
-        if (isDeleted) {
-            assertThat(mirrorContract.getObtainerId())
-                    .isEqualTo(contractClient
-                            .getSdkClient()
-                            .getExpandedOperatorAccountId()
-                            .getAccountId()
-                            .toString());
-        } else {
-            assertThat(mirrorContract.getObtainerId()).isNull();
-        }
-
         return mirrorContract;
-    }
-
-    private void verifyContractExecutionResultsById() {
-        List<ContractResult> contractResults = mirrorClient
-                .getContractResultsById(deployedParentContract.contractId().toString())
-                .getResults();
-
-        assertThat(contractResults).isNotEmpty().allSatisfy(this::verifyContractExecutionResults);
-    }
-
-    private void verifyContractExecutionResultsByTransactionId() {
-        ContractResult contractResult = mirrorClient.getContractResultByTransactionId(
-                networkTransactionResponse.getTransactionIdStringNoCheckSum());
-
-        verifyContractExecutionResults(contractResult);
-        assertThat(contractResult.getBlockHash()).isNotBlank();
-        assertThat(contractResult.getBlockNumber()).isPositive();
-        assertThat(contractResult.getHash()).isNotBlank();
     }
 
     private boolean isEmptyHex(String hexString) {
         return !StringUtils.hasLength(hexString) || hexString.equals("0x");
     }
 
-    private void verifyContractExecutionResults(ContractResult contractResult) {
+    protected void verifyContractExecutionResults(ContractResult contractResult) {
+        super.verifyContractExecutionResults(contractResult);
+
         ContractExecutionStage contractExecutionStage = isEmptyHex(contractResult.getFunctionParameters())
                 ? ContractExecutionStage.CREATION
                 : ContractExecutionStage.CALL;
 
-        assertThat(contractResult.getCallResult()).isNotBlank();
-        assertThat(contractResult.getContractId())
-                .isEqualTo(deployedParentContract.contractId().toString());
-        var createdIds = contractResult.getCreatedContractIds();
-        assertThat(createdIds).isNotEmpty();
-        assertThat(contractResult.getErrorMessage()).isBlank();
-        assertThat(contractResult.getFailedInitcode()).isBlank();
         assertThat(contractResult.getFrom())
                 .isEqualTo(FeatureInputHandler.evmAddress(contractClient
                         .getSdkClient()
                         .getExpandedOperatorAccountId()
                         .getAccountId()));
-        assertThat(contractResult.getGasLimit())
-                .isEqualTo(contractClient
-                        .getSdkClient()
-                        .getAcceptanceTestProperties()
-                        .getFeatureProperties()
-                        .getMaxContractFunctionGas());
-        assertThat(contractResult.getGasUsed()).isPositive();
-        assertThat(contractResult.getTo())
-                .isEqualTo(FeatureInputHandler.evmAddress(deployedParentContract.contractId()));
+
+        var createdIds = contractResult.getCreatedContractIds();
+        assertThat(createdIds).isNotEmpty();
 
         int amount = 0; // no payment in contract construction phase
         int numCreatedIds = 2; // parent and child contract
